Refactored out the commands into its own module
